refactor: rearrange app modules and utilities for better consistency
