+++ resources/client/modules/axios.ts
@@ -2,20 +2,6 @@ import { captureException } from '@sentry/vue'
 import axios from 'axios'
 import type { AxiosError, AxiosStatic } from 'axios'
 
-declare global {
-  const axios: AxiosStatic
-
-  interface Window {
-    axios: AxiosStatic
-  }
-}
-
-declare module 'vue' {
-  interface ComponentCustomProperties {
-    $axios: AxiosStatic
-  }
-}
-
 axios.defaults.headers.common['X-Requested-With'] = 'XMLHttpRequest'
 axios.defaults.withCredentials = true
 axios.interceptors.response.use(response => response, (error: AxiosError) => {
@@ -38,4 +24,18 @@ export const install: AppModuleInstall = ({ app, isClient }): void => {
     return
 
   app.config.globalProperties.$axios = window.axios = axios
+}
+
+declare global {
+  const axios: AxiosStatic
+
+  interface Window {
+    axios: AxiosStatic
+  }
+}
+
+declare module 'vue' {
+  interface ComponentCustomProperties {
+    $axios: AxiosStatic
+  }
 }
