+++ resources/client/shim.d.ts
@@ -17,6 +17,20 @@ interface UserProp {
   email: string
 }
 
+declare global {
+  type AppLocale = 'id' | 'en'
+
+  /**
+   * Application module install function.
+   */
+  type AppModuleInstall = (ctx: AppModuleContext) => void
+
+  interface AppModuleContext {
+    app: App<Element>
+    isClient: boolean
+  }
+}
+
 declare module '@inertiajs/vue3' {
   export function usePage<T extends AppPageProps>(): Page<T>
 }
